+++ examples/angular2/app/src/services/database.service.ts
@@ -77,7 +77,7 @@ export class DatabaseService {
         collections
             .filter(col => col.sync)
             .map(col => col.name)
-            .forEach(colName => db[colName].sync(syncURL + colName + '/'));
+            .forEach(colName => db[colName].sync({ remote: syncURL + colName + '/' }));
 
         return db;
     }