+++ examples/vue/src/database/Database.js
@@ -58,7 +58,9 @@ const _create = async function() {
 
     // sync
     console.log('DatabaseService: sync');
-    collections.filter(col => col.sync).map(col => col.name).map(colName => db[colName].sync(syncURL + colName + '/'));
+    collections.filter(col => col.sync).map(col => col.name).map(colName => db[colName].sync({
+        remote: syncURL + colName + '/'
+    }));
 
     return db;
 };